WP-1.1: money to minor units
- Migration 008: DECIMAL→BIGINT via round(amount*100) on chores, ledger_entries, settlements; paired down reversal
- Go: all money float64→int64 across models, handlers (chores/ledger/settlements), db repos
- Ledger balance query: add ::bigint casts on SUM(bigint) columns to prevent numeric scan failure
- openapi.yaml: amount/balance fields number/double→integer int64 with paise descriptions
- FE: regenerate api-types.gen.ts; remove rupeesToMinor shim + all call sites
- FE: add parseMoneyToMinorUnits + formatMinorUnits helpers in money.ts with unit tests
- FE: wire parseMoneyToMinorUnits into chores + ledger forms (replace parseFloat)
- FE: fix all toFixed(2)/$-prefixed displays in ledger/dashboard/group-overview to use formatMinorUnits
- CI: add -p 1 to integration test step to fix schema_migrations race flake

+++ backend/internal/db/settlement_repo.go
@@ -22,7 +22,7 @@ func NewSettlementRepo(pool *pgxpool.Pool) *SettlementRepo {
 }
 
 // Create inserts a new settlement
-func (r *SettlementRepo) Create(ctx context.Context, groupID, userID uuid.UUID, amount float64, date time.Time, note *string) (*models.Settlement, error) {
+func (r *SettlementRepo) Create(ctx context.Context, groupID, userID uuid.UUID, amount int64, date time.Time, note *string) (*models.Settlement, error) {
 	settlement := &models.Settlement{
 		ID:      uuid.New(),
 		GroupID: groupID,
